vim: Add complex types to syntax highlighting
This simply adds extra options into the type-keyword and designator
matching.  This means that there are now more things that would be
invalid QASM 3 programs, but will be syntax highlighted, but it isn't
really the job of the syntax highlighter to perfectly parse a program;
speed is preferable to total accuracy in negative matches.